feat: storybook - documentatie structuur aangepast
