+++ apps/storybook-e2e/src/e2e/ontwerp/popover-menu-accordion/popover-menu-accordion.cy.ts
@@ -1,0 +1,9 @@
+const PopoverMenuAccordionDefaultUrl =
+    'http://localhost:8080/iframe.html?args=&id=ontwerp-popover-menu-in-accordion--popover-menu-accordion&viewMode=story';
+
+describe('story popover-menu-accordion', () => {
+    it('should display story', () => {
+        cy.visit(PopoverMenuAccordionDefaultUrl);
+        cy.get('vl-accordion').shadow();
+    });
+});
