+++ apps/storybook/.storybook/main.ts
@@ -18,7 +18,6 @@ const config: StorybookConfig = {
     stories: [
         '../docs/**/*.stories.mdx',
         '../docs/**/*.stories.@(js|jsx|ts|tsx)',
-        '../docs/**/*.stories-doc.mdx',
         '../../../libs/components/src/**/*.stories.@(js|jsx|ts|tsx)',
         '../../../libs/elements/src/**/*.stories.@(js|jsx|ts|tsx)',
         '../../../libs/form/src/**/*.stories.@(js|jsx|ts|tsx)',
