--- apps/storybook/docs/d_applicatief/b_testen/1_klassieke-test-piramide.stories.mdx
@@ -1,24 +1,0 @@
-import {Meta} from '@storybook/addon-docs';
-import klassiekeTestPiramide from '../../../resources/applicatief-testen/klassieke-test-piramide.png';
-
-<Meta title="Applicatief/Testen/Klassieke Test Piramide"/>
-
-# Klassieke Test Piramide
-
-<br/>
-<img src={klassiekeTestPiramide} alt="Klassieke Test Piramide" height="350px"/>
-<br/>
-<br/>
-
-De meest simpele variant van de test piramide gaat op voor alle type toepassingen. Het is het minimale wat dient te
-gebeuren rond testen, de feitelijke uitvoering ligt open. Volgende concepten zitten vervat in de piramide:
-
-- het doel is minimale overlap te hebben tussen de niveaus: wat al getest is op een lager niveau dient niet meer getest
-  te worden op een hoger niveau
-- naarmate je dichter bij de top komt
-  * stijgt de complexiteit van de testen
-  * stijgt de implementatie tijd
-  * stijgt de doorlooptijd van uitvoering
-  * daalt de hoeveelheid van 'dat' type testen
-
-> **conceptueel: probeer 'wat' je wil testen zo laag als mogelijk in de piramide te doen**
