+++ apps/storybook/docs/d_richtlijnen/a_accessibility/1_accessibility.stories.mdx
@@ -1,0 +1,13 @@
+import {Meta} from '@storybook/addon-docs';
+
+<Meta title="Richtlijnen/Accessibility/Accessibility"/>
+
+# Accessibility
+
+## Referenties
+
+[Digitale Toegankelijkheid - Vlaanderen.be](https://www.vlaanderen.be/inter/toolbox-toegankelijke-steden-en-gemeenten/algemeen-bestuur-dienstverlening-en-communicatie/digitale-toegankelijkheid)
+
+[W3 Accessibility](https://www.w3.org/TR/WCAG21/#:~:text=Web%20Content%20Accessibility%20Guidelines%20(WCAG)%202.1%20defines%20how%20to%20make,%2C%20learning%2C%20and%20neurological%20disabilities)
+
+[W3 WAI](https://www.w3.org/WAI/)
