+++ apps/storybook/docs/d_richtlijnen/b_testen/5_selenium-en-cypress.stories.mdx
@@ -1,0 +1,65 @@
+import {Meta} from '@storybook/addon-docs';
+
+<Meta title="Richtlijnen/Testen/Selenium en Cypress"/>
+
+# Selenium en Cypress
+
+## Inleiding
+
+Bij Departement Omgeving wordt typisch Selenium gebruikt voor frontend testen. Selenium wordt algemeen veel gebruikt,
+is goed gedocumenteerd, maar niet ideaal. Het artikel
+[cypress, the selenium killer](https://blog.logrocket.com/cypress-io-the-selenium-killer/) beschrijft de problemen die
+ik typisch ook ervaar met Selenium.
+
+
+## Selenium
+
+Een aantal algemene artikels die Selenium goed beschrijven zijn:
+
+- de [Selenium website](https://www.selenium.dev)
+- de [Selenium wikipedia](https://en.wikipedia.org/wiki/Selenium_(software)) pagina
+- een [introductie tot Selenium](https://www.guru99.com/introduction-to-selenium.html)
+
+Selenium heeft een [flexibele architectuur](https://www.interviewbit.com/blog/selenium-architecture) met web-driver's
+waardoor het historisch een breed scala van talen (om testen in te schrijven) en browsers ondersteund. Typisch werden
+deze testen (vroeger in [SSR toepassingen](/docs/applicatief-testen-test-piramide-ssr-vs-spa--documentatie)) in dezelfde taal als de
+applicatie (bij DOMG dus in Java) geschreven. Toepassingen zijn geëvolueerd richting SPA's, in JavaScript geschreven.
+Daar Selenium ook JavaScript ondersteuning heeft is dan vaak die test tool behouden. In de opzet van een SPA zijn er
+echter betere alternatieven.
+
+
+## Eenduidigheid
+
+Volgende factoren maken het minder eenduidig maken wat er in de DOM gebeurd:
+
+- bij een responsive toepassing zullen afhankelijk van de resolutie waarop getest wordt elementen onzichtbaar worden,
+  een ander formaat of andere positie hebben
+- door web-componenten met een shadow DOM is de DOM-structuur minder transparant
+- frameworks en component bibliotheken doen soms auto-magische zaken
+
+
+## Selenium en SPA's
+
+### Page Object Model
+
+De [POM](https://www.selenium.dev/documentation/test_practices/encouraged/page_object_models/) aanpak die vaak
+gebruikt wordt is zinvol voor SSR's, je hebt verschillende pagina's met hun eigen url. Er is een 1 op 1 mapping
+van de pagina met het POM. Voor een SPA is dat lastiger, er is maar 1 page waarbinnen alles dynamisch gebeurd.
+Vasthouden aan het POM kan, maar zal niet eenduidig zijn.
+
+
+## Cypress
+
+[Cypress](https://www.cypress.io/)
+https://www.perfecto.io/blog/cypress-vs-selenium-whats-right-cross-browser-testing-solution-you
+https://www.browserstack.com/guide/cypress-vs-selenium
+https://www.blazemeter.com/blog/selenium-vs-cypress
+
+
+## OpenWC Testing
+
+[web-component-tester](https://www.npmjs.com/package/web-component-tester)
+[OpenWC Testing](https://open-wc.org/docs/testing/testing-package/)
+
+
+## Storybook
