+++ libs/form/src/next/checkbox/stories/vl-checkbox.stories-doc.mdx
@@ -7,7 +7,7 @@ import FormControlPublicMethods from '../../form-control/stories/form-control.pu
 Gebruik de `checkbox-next` component om de gebruiker de mogelijkheid te geven om 1 of meerdere dingen te selecteren in een lijst.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-checkbox](/docs/components-checkbox--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-checkbox](/docs/components-checkbox--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
