+++ libs/form/src/next/datepicker/stories/vl-datepicker.stories-doc.mdx
@@ -7,7 +7,7 @@ import FormControlPublicMethods from '../../form-control/stories/form-control.pu
 Gebruik de `datepicker-next` component om de gebruiker op een gebruiksvriendelijke manier een datum of tijd te laten selecteren.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-datepicker](/docs/components-datepicker--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-datepicker](/docs/components-datepicker--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
