+++ libs/form/src/next/error-message/stories/vl-error-message.stories-doc.mdx
@@ -6,7 +6,7 @@ import * as VlErrorMessageStories from './vl-error-message.stories';
 Gebruik de `error-message-next` component om een error af te beelden voor een input veld.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-form-validation-message](/docs/elements-form-message-form-validation-message--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-form-validation-message](/docs/elements-form-message-form-validation-message--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
