+++ libs/form/src/next/input-field/stories/vl-input-field.stories-doc.mdx
@@ -7,7 +7,7 @@ import FormControlPublicMethods from '../../form-control/stories/form-control.pu
 Gebruik de `input-field-next` component om een input veld toe te voegen aan een pagina.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-input-field](/docs/elements-input-field--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-input-field](/docs/elements-input-field--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
