+++ libs/form/src/next/radio-group/stories/vl-radio-group.stories-doc.mdx
@@ -9,7 +9,7 @@ Wanneer mogelijk, selecteer geen `radio` van vooraf zodat de gebruiker een bewus
 De `radio-next` component is een onderdeel van de `radio-group-next` component.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van het [vl-radio](/docs/components-radio--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van het [vl-radio](/docs/components-radio--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
