+++ libs/form/src/next/select/stories/vl-select.stories-doc.mdx
@@ -7,7 +7,7 @@ import FormControlPublicMethods from '../../form-control/stories/form-control.pu
 Gebruik de `select-next` component om een select of multiselect veld toe te voegen aan een pagina.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-select](/docs/elements-select--documentatie) component en de [vl-multiselect](/docs/elements-multiselect--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-select](/docs/elements-select--documentatie) component en de [vl-multiselect](/docs/elements-multiselect--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
