+++ libs/form/src/next/textarea/stories/vl-textarea.stories-doc.mdx
@@ -7,7 +7,7 @@ import FormControlPublicMethods from '../../form-control/stories/form-control.pu
 Gebruik de `textarea-next` component om een textarea veld toe te voegen aan een pagina.<br/>
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-textarea](/docs/components-textarea--documentatie) component, we raden aan deze versie te gebruiken.<br />
+Dit component is de volgende versie van de [vl-textarea](/docs/components-textarea--documentatie) component, we raden aan op termijn deze versie te gebruiken.<br />
 In tegenstelling tot de vorige [vl-textarea](/docs/components-textarea--documentatie) bevat dit component geen rich-text modus, hier wordt op termijn een ander component voor voorzien.
 
 ## Voorbeeld
