+++ libs/form/src/next/upload/stories/vl-upload.stories-doc.mdx
@@ -8,7 +8,7 @@ Gebruik de `upload-next` component om één of meerdere bestanden te selecteren 
 
 Zie het [form demo](/docs/applicatief-voorbeelden-form-demo--documentatie) voorbeeld voor het gebruik binnen een form.
 
-Dit component is de volgende versie van de [vl-upload](/docs/components-upload--documentatie) component, we raden aan deze versie te gebruiken.
+Dit component is de volgende versie van de [vl-upload](/docs/components-upload--documentatie) component, we raden aan op termijn deze versie te gebruiken.
 
 ## Voorbeeld
